Convert MEME to TEME before generating TLEs; archive ephemeris data to S3 (disabled)
